dvr: tuner framework modify for dvr [1/1]
PD#SWPL-101150

Problem:
make adaption for dvr under tuner framework

Solution:
make adaption for dvr under tuner framework

Verify:
ohm

+++ logicdtvkit/src/logictuner/tuner_jni/dvb_frontend_setting_utils.cpp
@@ -89,7 +89,7 @@ jobject dvb_utils_getDvbcFrontendSettingsObject(JNIEnv *env, Dvbc_Frontend_Setti
 
     if (ANDROID_SDK_R < Am_tuner_getAndroidVersion()) {
         dvbcSettingInit = env->GetMethodID(dvbcSettingClass, "<init>", "(JIJIIIIII)V");
-        dvbcSettingObject = env->NewObject(dvbcSettingClass, dvbcSettingInit, (jlong)frequency, modulation, innerFec, symbolRate,
+        dvbcSettingObject = env->NewObject(dvbcSettingClass, dvbcSettingInit, (jlong)frequency, modulation, (jlong)innerFec, symbolRate,
                                                outerFec, annex, spectralInversion, 0, 0);
 
     } else {
